+++ app/src/main/java/io/nekohasekai/sagernet/widget/UserAgentPreference.kt
@@ -26,6 +26,7 @@ class UserAgentPreference
     companion object {
         val PRESETS = listOf(
             "默认 (使用全局默认)",
+            "NekoBox/Android/1.4.2 (Prefer ClashMeta Format)",
             "sing-box/1.14",
             "sing-box/1.14.0",
             "ClashMeta",